Try out smaller 8-bit version on fix_fft

+++ firmware/fft/fix_fft.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-int fix_fft(int16_t fr[], int16_t fi[], int16_t m, int16_t inverse);
-int fix_fftr(int16_t f[], int m, int inverse);
+int fix_fft(int8_t fr[], int8_t fi[], int m, int inverse);
+int fix_fftr(int8_t f[], int m, int inverse);
 
 #endif
